clean up pages and components, add menu, tweak navbar
