+++ gui/src/Tests.js
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom';
 // import { Button, Icon, Menu, Segment } from 'semantic-ui-react';
 import { Button } from 'semantic-ui-react';
 import './style.css';
-import { Header } from './Header.js';
+import { Header } from './components/Header.js';
 import { apiUrl } from './utils.js';
 
 export function TestButton(props) {
@@ -39,7 +39,7 @@ export function Tests() {
 		<div>
 			<Header>Hardware Tests</Header>
 
-			<div className='button-menu'>
+			<div className='button-list'>
 				<Button as={Link} to='/relays'>
 					Relays
 				</Button>
