+++ gui/src/pages/Settings.js
@@ -1,13 +1,9 @@
 import React from 'react';
 // import { Button, Icon, Menu, Segment } from 'semantic-ui-react';
 
-import { Header } from './Header.js';
-
 export function Settings() {
 	return (
 		<div>
-			<Header>Settings</Header>
-
 			<p>Settings page here...</p>
 		</div>
 	);
